pyproject.toml
[project]
name = "ccgram"
dynamic = ["version"]
description = "CCGram — manage AI coding agents from Telegram via tmux"
readme = "README.md"
license = "MIT"
requires-python = ">=3.14"
authors = [
    { name = "Alexei Ledenev", email = "alexei.led@gmail.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.14",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development",
    "Typing :: Typed",
]
dependencies = [
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "libtmux>=0.50.0",
    "Pillow>=10.0.0",
    "telegramify-markdown>=1.0.0",
    "aiofiles>=24.0.0",
    "structlog>=24.0.0",
    "click>=8.1.0",
    "pyte>=0.8.2",
    "openai>=1.58.0",
]

[project.urls]
Homepage = "https://github.com/alexei-led/ccgram"
Repository = "https://github.com/alexei-led/ccgram"
Issues = "https://github.com/alexei-led/ccgram/issues"
Changelog = "https://github.com/alexei-led/ccgram/blob/main/CHANGELOG.md"

[project.scripts]
ccgram = "ccgram.main:main"

[project.optional-dependencies]
dev = [
    "deptry>=0.23.0",
    "pyright>=1.1.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0",
    "pytest-timeout>=2.3",
    "ruff>=0.8.0",
    "hypothesis>=6.0",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
local_scheme = "no-local-version"

[tool.hatch.build.hooks.vcs]
version-file = "src/ccgram/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/ccgram"]

[tool.ruff]
target-version = "py314"
exclude = ["src/ccgram/_version.py"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "ARG", "G", "BLE", "SIM", "PLR2004", "C901", "PLR0911", "PLR0912", "PLR0915", "N"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pylint]
max-returns = 8
max-branches = 15
max-statements = 60

[tool.ruff.lint.per-file-ignores]
# Tests: unused args (fixtures, mock lambdas), magic values in assertions
"tests/**" = ["ARG", "PLR2004"]
# ANSI SGR codes/Unicode ranges + ANSI code parsing is inherently branchy
"src/ccgram/screenshot.py" = ["PLR2004", "C901", "PLR0912"]
# Transcript parser: inherently complex JSONL parsing and tool dispatch
"src/ccgram/transcript_parser.py" = ["C901", "PLR0912", "PLR0915", "PLR0911"]
# Session manager: complex state migration (resolve_stale_ids) and session map loading
"src/ccgram/session.py" = ["C901", "PLR0912", "PLR0915"]
# Session monitor: filesystem scan with many validation branches
"src/ccgram/session_monitor.py" = ["C901", "PLR0912"]
# Bot: callback dispatch and message routing are inherently branchy
"src/ccgram/bot.py" = ["C901", "PLR0912"]
# Hook: CLI dispatch with multiple subcommands
"src/ccgram/hook.py" = ["C901", "PLR0915"]
# CC commands: filesystem scan with validation branches
"src/ccgram/cc_commands.py" = ["C901", "PLR0912"]
# History: paginated message formatting with many edge cases
"src/ccgram/handlers/history.py" = ["C901", "PLR0912", "PLR0915"]
# Response builder: message formatting with content-type dispatch
"src/ccgram/handlers/response_builder.py" = ["C901", "PLR0912"]
# Status polling: background loop with multiple state checks
"src/ccgram/handlers/status_polling.py" = ["C901", "PLR0912"]
# Message queue: worker loop with task dispatch and error handling
"src/ccgram/handlers/message_queue.py" = ["C901"]
# Directory callbacks: confirm handler with state cleanup branches
"src/ccgram/handlers/directory_callbacks.py" = ["C901"]
# Recovery callbacks: fresh handler with state cleanup branches
"src/ccgram/handlers/recovery_callbacks.py" = ["C901"]

[tool.deptry.per_rule_ignores]
# DEP002: dev/CLI tools (never imported in src/) + httpx (transitive runtime dep of python-telegram-bot)
DEP002 = ["httpx", "deptry", "pyright", "pytest", "pytest-asyncio", "pytest-cov", "pytest-timeout", "ruff", "hypothesis"]

[tool.pyright]
pythonVersion = "3.14"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: tests using real filesystem or external processes",
    "e2e: end-to-end tests with real agent CLIs (local only, slow)",
]

[tool.coverage.run]
source = ["ccgram"]
branch = true

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "logger\\.",
]
